quite a few changes and bug fixes; closes issue #4; moved to version 0.4.0

+++ text_importer/importer.py
@@ -2,16 +2,17 @@
 Functions and CLI script to convert Olive OCR data into Impresso's format.
 
 Usage:
-    import.py --input-dir=<id> --output-dir==<od> [--log-file=<f> --temp-dir==<td> --verbose --parallelize --filter=<ft>]
+    import.py --input-dir=<id> [--output-dir==<od> --s3-bucket=<b> --log-file=<f> --temp-dir==<td> --verbose --parallelize --filter=<ft>]
 
 Options:
     --input-dir=<id>    Base directory containing one sub-directory for each journal.
-    --output-dir==<od>  Base directory where to write the output files.
+    --output-dir=<od>   Base directory where to write the output files.
+    --s3-bucket=<b>     If provided, writes output to an S3 drive, in the specified bucket.
     --log-file=<f>      Log file; when missing print log to stdout
     --verbose           Verbose log messages (good for debugging).
     --parallelize       Parallelize the import.
     --filter=<ft>       Criteria to filter issues before import ("journal=GDL; date=1900/01/01-1950/12/31;")
-"""
+"""  # noqa: E501
 
 
 import json
@@ -35,7 +36,6 @@ __email__ = "matteo.romanello@epfl.ch"
 __organisation__ = "impresso @ DH Lab, EPFL"
 __copyright__ = "EPFL, 2017"
 __status__ = "development"
-__version__ = "0.3.0"
 
 logger = logging.getLogger()
 
@@ -51,19 +51,6 @@ html_escape_table = {
     "&gt;": ">",
     "&lt;": "<",
 }
-
-
-def print_article(article):
-    """Only for debug, remove later."""
-    print(article["legacy"]["id"])
-    for r in article["regions"]:
-        for p in r["paragraphs"]:
-            print("------------")
-            assert p["lines"] is not None
-            for line in p["lines"]:
-                print(" ".join(t['text'] for t in line["tokens"]))
-            print("------------")
-        print("#############")
 
 
 def _parse_filter(filter_string):
@@ -120,6 +107,7 @@ def main(args):
     # store CLI parameters
     inp_dir = args["--input-dir"]
     outp_dir = args["--output-dir"]
+    out_bucket = args["--s3-bucket"]
     temp_dir = args["--temp-dir"]
     log_file = args["--log-file"]
     parallel_execution = args["--parallelize"]
@@ -144,7 +132,7 @@ def main(args):
     logger.debug("CLI arguments received: {}".format(args))
 
     # clean output directory if existing
-    if os.path.exists(outp_dir):
+    if outp_dir is not None and os.path.exists(outp_dir):
         shutil.rmtree(outp_dir)
 
     # clean temp directory if existing
@@ -176,10 +164,19 @@ def main(args):
 
     logger.debug("Following issues will be imported:{}".format(issues))
 
-    result = [
-        olive_import_issue(i, outp_dir, temp_dir)
-        for i in issues
-    ]
+    assert outp_dir is not None or out_bucket is not None
+
+    if outp_dir is not None:
+        result = [
+            olive_import_issue(i, out_dir=outp_dir, temp_dir=temp_dir)
+            for i in issues
+        ]
+    elif out_bucket is not None:
+        result = [
+            olive_import_issue(i, s3_bucket=out_bucket, temp_dir=temp_dir)
+            for i in issues
+        ]
+
     """
     # prepare the execution of the import function
     tasks = [
